rt: reorganize I/O driver source
Moves `Driver` into its own file and eliminates a bunch of code defined
in macros.
