+++ tokio/src/runtime/io/registration.rs
@@ -199,6 +199,33 @@ impl Registration {
         }
     }
 
+    pub(crate) async fn readiness(&self, interest: Interest) -> io::Result<ReadyEvent> {
+        let ev = self.shared.readiness(interest).await;
+
+        if ev.is_shutdown {
+            return Err(gone());
+        }
+
+        Ok(ev)
+    }
+
+    pub(crate) async fn async_io<R>(
+        &self,
+        interest: Interest,
+        mut f: impl FnMut() -> io::Result<R>,
+    ) -> io::Result<R> {
+        loop {
+            let event = self.readiness(interest).await?;
+
+            match f() {
+                Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
+                    self.clear_readiness(event);
+                }
+                x => return x,
+            }
+        }
+    }
+
     fn handle(&self) -> &Handle {
         self.handle.driver().io()
     }
@@ -222,31 +249,4 @@ fn gone() -> io::Error {
         io::ErrorKind::Other,
         crate::util::error::RUNTIME_SHUTTING_DOWN_ERROR,
     )
-}
-
-cfg_io_readiness! {
-    impl Registration {
-        pub(crate) async fn readiness(&self, interest: Interest) -> io::Result<ReadyEvent> {
-            let ev = self.shared.readiness(interest).await;
-
-            if ev.is_shutdown {
-                return Err(gone())
-            }
-
-            Ok(ev)
-        }
-
-        pub(crate) async fn async_io<R>(&self, interest: Interest, mut f: impl FnMut() -> io::Result<R>) -> io::Result<R> {
-            loop {
-                let event = self.readiness(interest).await?;
-
-                match f() {
-                    Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
-                        self.clear_readiness(event);
-                    }
-                    x => return x,
-                }
-            }
-        }
-    }
 }
